back/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
sqlmodel>=0.0.16
httpx>=0.27
python-dotenv>=1.0
pytest>=8.0
eval_type_backport>=0.2; python_version < "3.10"